refactor: move tenable pipeline into function
